requirements.txt
# ============================================
# Core Framework
# ============================================
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# ============================================
# Async Task Queue
# ============================================
celery[redis]==5.3.6
redis==5.0.1

# ============================================
# Speech-to-Text
# ============================================
openai-whisper
torch>=2.0.0
torchaudio>=2.0.0

# ============================================
# LLM Providers
# ============================================
# Anthropic (Claude)
anthropic==0.18.1

# OpenAI (GPT-4)
openai==1.12.0

# Google Gemini
google-generativeai==0.4.1

# Groq
groq==0.4.2

# ============================================
# Audio Processing
# ============================================
pydub==0.25.1
ffmpeg-python==0.2.0

# ============================================
# HTTP & Async
# ============================================
httpx==0.27.0
aiofiles==23.2.1

# ============================================
# Data Validation
# ============================================
pydantic==2.6.1
pydantic-settings==2.1.0

# ============================================
# Utilities
# ============================================
python-dotenv==1.0.1
tenacity==8.2.3

# ============================================
# Database (Optional)
# ============================================
sqlalchemy==2.0.25
aiosqlite==0.19.0

# ============================================
# Monitoring (Optional)
# ============================================
flower==2.0.1
